+++ CHANGES.md
@@ -1,1 +1,4 @@
 # Changelog
+
+## implement multilingual greet function
+- Implemented at 2026-06-25T17:56:36Z